Ensure a valid CA cert is written to relations
Early during the unit lifecycle before consuming applications
present requests its quite possible that the string intepretation
of the Python value None can get written to the relation for the
ca and chain key.

Guard against this - consuming units will inteprete this as a
string rather than being unset which will then cause hook errors
when the charm parses the "None" string as a certificate.